requirements.txt
streamlit
python-dotenv
camel-ai
requests
arxiv
astor
arxiv2text
mistralai
